refactor: Snackbar 관련 컴포넌트 패키지 위치 이동 (`modal` 서브 패키지 추가)
